requirements.txt
python-dotenv
langchain_groq
langchain
pydantic
PySide6